requirements.txt
asgiref==3.3.4
biopython==1.79
dj-database-url==0.5.0
Django==3.2.4
django-background-tasks==1.2.5
django-compat==1.0.15
gunicorn==20.1.0
numpy==1.20.3
psycopg2==2.8.6
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
whitenoise==5.2.0
